legacy-python/requirements.txt
pyjwt
requests
msal
python-dotenv
pyodbc
pymilvus
neo4j
transformers
torch
librosa
Pillow
pypdf
numpy
PyPDF2
pytesseract
SpeechRecognition
marshmallow
sentence-transformers
tf-keras
black
isort
ruff
opencensus-ext-azure
azure-keyvault-secrets
aioodbc
gunicorn
pytest-asyncio
fastapi
uvicorn[standard]
pydantic
aiofiles
google-generativeai
anthropic
huggingface_hub
httpx